backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_groq
requests
pydantic
google-cloud-speech
google-cloud-texttospeech
langchain-huggingface
langchain-chroma
sentence-transformers
torch
transformers
nest-asyncio   
openai-whisper
ffmpeg-python
python-multipart